Make CarouselRoot a nested class
