+++ osu.Game/Screens/Select/BeatmapCarousel.cs
@@ -628,5 +628,23 @@ namespace osu.Game.Screens.Select
             // layer transformations on top, with a similar reasoning to the previous comment.
             p.SetMultiplicativeAlpha(MathHelper.Clamp(1.75f - 1.5f * dist, 0, 1));
         }
+
+        private class CarouselRoot : CarouselGroupEagerSelect
+        {
+            private readonly BeatmapCarousel carousel;
+
+            public CarouselRoot(BeatmapCarousel carousel)
+            {
+                this.carousel = carousel;
+            }
+
+            protected override void PerformSelection()
+            {
+                if (LastSelected == null)
+                    carousel.SelectNextRandom();
+                else
+                    base.PerformSelection();
+            }
+        }
     }
 }
